\name{ERgadem.15FDR}
\alias{ERgadem.15FDR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rGADEM object obtained from the 15\% FDR sequences detected by rMAT on the ER data.
}
\description{
motifs GADEM 
}

\usage{
  data(ERgadem)
}

\value{
Loading the data returns an object named ERgadem.15FDR of class GADEM.
}

\source{

}

\references{

}

\author{
Raphael Gottardo
}

\seealso{
\code{\link{GADEM}}
}
\examples{
data(ERgadem.15FDR)
summary(ERgadem.15FDR)
nMotifs(ERgadem.15FDR)
nOccurrences(ERgadem.15FDR)
}
