\name{ERgadem}
\alias{ERgadem}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rGADEM object obtained from the 10\% FDR sequences detected by PICS with fixed length 200 on the ER data.
}
\description{
motifs GADEM 
}

\usage{
  data(ERgadem)
}

\value{
Loading the data returns an object named ERgadem of class GADEM.
}

\source{

}

\references{

}

\author{
Raphael Gottardo
}

\seealso{
\code{\link{GADEM}}
}
\examples{
data(ERgadem)
summary(ERgadem)
nMotifs(ERgadem)
nOccurrences(ERgadem)
}
