\name{FOXA1gadem}
\alias{FOXA1gadem}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rGADEM object obtained from the 10\% FDR sequences detected by PICS with fixed length 200 on the FOXA1 data.
}
\description{
motifs GADEM 
}

\usage{
  data(FOXA1gadem)
}

\value{
Loading the data returns an object named FOXA1gadem of class GADEM.
}

\source{

}

\references{

}

\author{
Raphael Gottardo
}

\seealso{
\code{\link{GADEM}}
}
\examples{
data(FOXA1gadem)
summary(FOXA1gadem)
nMotifs(FOXA1gadem)
nOccurrences(FOXA1gadem)
}
